introduce defer functionality
Exception safety in Python can easily pollute code with a lot of
try-except-finally blocks and can be quite tricky to achieve. This is
not exactly alleviated by the fact that documentation about which
function might raise which exceptions (if any) is scarce, if not
non-existant.
This change introduces a means helping to get confidence in the cleanup
code. The idea is to create a "context" of some sort that is known to
get destroyed properly (by using Python's 'with' blocks). Whenever
automated cleanup is desired, clients can register functions with this
context that are executed unconditionally whenever the context is
destroyed, i.e., the with block exited.
The function to create a context is called defer, and usage could look
like this:
  client = Client()
  with defer() as d:
    obj = Object()
    d.defer(obj.destroy)
    obj.register(client)
    d.defer(lambda: obj.unregister(client))
    raise Exception()

In this example, although an exception is raised, the client is
unregistered and the object is destroyed properly. This example also
illustrates an important property: deferred functions are invoked in
reverse order on cleanup.