Introduce release functionality to defer module
The defer module can help in safely rolling back previously made changes
in case of an error (that is, an exception being raised). Until now,
once a function was registered to be deferred it was executed
eventually. Sometimes, however, this behavior is not desired. Consider
the case where an object is initialized in various steps. Each step
could be guarded be a defer operation and would be reverted in case of
error. If the initialization succeeds, however, the deferred functions
would be invoked unconditionally, causing the object to be torn down
again. Because of this property using the defer functionality in such
scenarios is not rewarding.
With this change we can support such scenarios: the newly introduced
release functionality allows for a per function or per defer object
unregistration of roll backs. In the previously mentioned scenario a
single release() call on the defer context would void all reversal
functions, keeping the successfully initialized state.