Allow for parameter passing to deferred function
Until now a deferred function had to be directly executable, without the
need for any additional arguments. If a function takes additional
arguments, it has to be wrapped into a lambda expression. In principle,
this behavior is fine. However, argument binding to lambda functions can
be tricky. Consider the following example:
  with defer() as d:
    for i in range(3):
      d.defer(lambda: print(i))

When the with block is exited, the output will be three times the number
two instead of the intuitively expected sequence zero, one, and two. The
reason is rooted in Python's variable binding rules. In a couple of
occasions this behavior makes it necessary to define additional
functions. The defer functionality is prone to these problems when a
function requiring parameters is to be deferred.
With this patch we add direct parameter passing capabilities to the defer
function of a defer context object. This change alleviates the need for
defining a lambda expression in the calling scope and, hence, avoids the
aforementioned problem. It also makes function deferral more concise while
being fully compatible with the previous behavior.